+++ client/src/components/SearchWorld.js
@@ -1,9 +1,7 @@
 import React, {Component} from 'react';
 import { View, Image, Dimensions, Text, StatusBar, TextInput, TouchableOpacity } from 'react-native';
 import { Actions } from 'react-native-router-flux';
-import Carousel from 'react-native-snap-carousel';
-
-import HotLocation from './HotLocation';
+import Carousel from 'react-native-looped-carousel';
 
 const { height, width } = Dimensions.get('window');
 
@@ -18,13 +16,7 @@ class SearchWorld extends Component {
     Actions.MapContainer({ ManualAddress:this.state.address });
     this.setState({ address: '' });
   }
-   _renderItem (entry) {
-    return (
-        <HotLocation {...entry} />
-    );
-  }
   render(){
-    
     return (
     <View>
       <StatusBar
@@ -62,21 +54,14 @@ class SearchWorld extends Component {
           </View>
           <View>
             <Text style={styles.locationHeaderText}>Hot locations (tap to select)</Text>
-            <Carousel
-              items={entries}
-              firstItem={2}
-              inactiveSlideScale={0.94}
-              inactiveSlideOpacity={0.6}
-              renderItem={this._renderItem}
-              sliderWidth={sliderWidth}
-              itemWidth={itemWidth}
-              slideStyle={styles.slide}
-              containerCustomStyle={{ marginBottom: 35 }}
-              //contentContainerCustomStyle={sliderStyles.sliderContainer}
-              showsHorizontalScrollIndicator={false}
-              snapOnAndroid={true}
-              removeClippedSubviews={false}
-            />
+            <Carousel style={{ alignSelf: 'center', height: 200, width: 200 }}>
+            <View style={styles.hotSpotView}>
+              <Text style={{color:'white', backgroundColor: 'transparent',}}>Hi</Text>
+            </View>
+            <View style={styles.hotSpotView}>
+              <Text style={{color:'white', backgroundColor: 'transparent',}}>Hey</Text>
+            </View>
+            </Carousel>
           </View>
         </View>
       </Image>
@@ -84,44 +69,6 @@ class SearchWorld extends Component {
     )
   }
 };
-const entries = [
-    {
-        title: 'Beautiful and dramatic Antelope Canyon',
-        subtitle: 'Lorem ipsum dolor sit amet et nuncat mergitur',
-        illustration: 'http://i.imgur.com/UYiroysl.jpg'
-    },
-    {
-        title: 'Earlier this morning, NYC',
-        subtitle: 'Lorem ipsum dolor sit amet',
-        illustration: 'http://i.imgur.com/UPrs1EWl.jpg'
-    },
-    {
-        title: 'White Pocket Sunset',
-        subtitle: 'Lorem ipsum dolor sit amet et nuncat ',
-        illustration: 'http://i.imgur.com/MABUbpDl.jpg'
-    },
-    {
-        title: 'Acrocorinth, Greece',
-        subtitle: 'Lorem ipsum dolor sit amet et nuncat mergitur',
-        illustration: 'http://i.imgur.com/KZsmUi2l.jpg'
-    },
-    {
-        title: 'The lone tree, majestic landscape of New Zealand',
-        subtitle: 'Lorem ipsum dolor sit amet',
-        illustration: 'http://i.imgur.com/2nCt3Sbl.jpg'
-    },
-    {
-        title: 'Middle Earth, Germany',
-        subtitle: 'Lorem ipsum dolor sit amet',
-        illustration: 'http://i.imgur.com/lceHsT6l.jpg'
-    }
-    ]
- function wp(percentage) {
-    const value = (percentage * viewportWidth) / 100;
-    return Math.round(value);
-  }
-const itemWidth = slideWidth + itemHorizontalMargin * 2;
-const slideWidth = wp(75);
 const styles = {
   navBar: {
     width,
@@ -170,21 +117,16 @@ const styles = {
   locationHeaderText: {
     backgroundColor: 'transparent',
     color: '#EFEFF4',
-    fontSize: 26,
+    fontSize: 18,
     textAlign: 'center'
   },
   hotSpotView: {
     borderColor: '#EFEFF4',
     borderWidth: 3,
     borderRadius: 5,
-    height: 250,
-    width: 250,
+    height: 100,
+    width: 100,
     alignSelf: 'center'
-  },
-  slide: {
-    flexDirection: 'column',
-    width: itemWidth,
-    paddingHorizontal: itemHorizontalMargin
   }
 };
 export default SearchWorld;
